+++ imgaug/imgaug.py
@@ -781,6 +781,54 @@ def quokka_bounding_boxes(size=None, extract=None):
     return bbsoi
 
 
+def quokka_polygons(size=None, extract=None):
+    """
+    Returns example polygons on the standard example quokke image.
+
+    The result contains one polygon, covering the quokka's outline.
+
+    Parameters
+    ----------
+    size : None or float or tuple of int or tuple of float, optional
+        Size of the output image on which the polygons are placed. If None,
+        then the polygons are not projected to any new size (positions on the
+        original image are used). Floats lead to relative size changes, ints
+        to absolute sizes in pixels.
+
+    extract : None or 'square' or tuple of number or imgaug.BoundingBox or \
+              imgaug.BoundingBoxesOnImage
+        Subarea to extract from the image. See :func:`imgaug.quokka`.
+
+    Returns
+    -------
+    psoi : imgaug.PolygonsOnImage
+        Example polygons on the quokka image.
+
+    """
+    left, top = 0, 0
+    if extract is not None:
+        bb_extract = _quokka_normalize_extract(extract)
+        left = bb_extract.x1
+        top = bb_extract.y1
+    with open(QUOKKA_ANNOTATIONS_FP, "r") as f:
+        json_dict = json.load(f)
+    polygons = []
+    for poly_json in json_dict["polygons"]:
+        polygons.append(
+            Polygon([(point["x"] - left, point["y"] - top)
+                    for point in poly_json["keypoints"]])
+        )
+    if extract is not None:
+        shape = (bb_extract.height, bb_extract.width, 3)
+    else:
+        shape = (643, 960, 3)
+    psoi = PolygonsOnImage(polygons, shape=shape)
+    if size is not None:
+        shape_resized = _compute_resized_shape(shape, size)
+        psoi = psoi.on(shape_resized)
+    return psoi
+
+
 def angle_between_vectors(v1, v2):
     """
     Returns the angle in radians between vectors `v1` and `v2`.